+++ src/config/constants.py
@@ -4,7 +4,7 @@ Application Constants
 
 APP_NAME = "Book Library Reconciler"
 
-VERSION = "2.0.0-alpha.3"
+VERSION = "2.0.0-alpha.4"
 
 # --------------------------------------------------
 # Display